upload

+++ parseXML.py
@@ -2,6 +2,7 @@
 
 import sys
 import math
+import json
 import xml.dom.minidom
 from pydub import AudioSegment
 
@@ -27,13 +28,17 @@ class NoteBase(object):
 class NoteN(NoteBase):
 	def __init__(self, note, tp):
 		super(NoteN, self).__init__(note, tp)
+		self.isSkill = False
 		
 	def generate(self, resultList):
 		resultList.append({
-			"type": "SingleOff" if isGrayNote(self.pos) else "Single",
+			"type": "Skill" if self.isSkill else ("SingleOff" if isGrayNote(self.pos) else "Single"),
 			"lane": self.line,
 			"time": calcTime(self.pos),
 		})
+		
+	def setToSkill(self):
+		self.isSkill = True
 
 # 粉键
 class NoteF(NoteBase):
@@ -55,6 +60,7 @@ class LineStart(NoteBase):
 		self.lineInsts = []
 		self.nextNode = None
 		self.prevNode = None
+		self.isSkill = False
 		
 	def addLineInst(self, nodeInst):
 		self.lineInsts.append(nodeInst)
@@ -73,7 +79,7 @@ class LineStart(NoteBase):
 		
 	def generate(self, resultList):
 		resultList.append({
-			"type": "Long",
+			"type": "Skill" if self.isSkill else "Long",
 			"lane": self.line,
 			"time": calcTime(self.pos),
 		})
@@ -82,6 +88,9 @@ class LineStart(NoteBase):
 			"lane": [self.line, self.nextNode.line],
 			"time": [calcTime(self.pos), calcTime(self.nextNode.pos)],
 		})
+		
+	def setToSkill(self):
+		self.isSkill = True
 
 # 滑条中间节点
 class LineMiddle(NoteBase):
@@ -149,15 +158,18 @@ bpmInfo = {}
 
 # 支持的命令行参数：
 # -sav super.sav 输入的sav文件名；
-# -map 导出map谱面，否则导出simulator谱面；
+# -map 导出map谱面；
+# -simulator 导出simulator谱面；
 # -mp3 super.mp3 输入的mp3文件名
 # -preLength 16 在谱面开始前添加的空白长度（以1/8拍的整数倍计算，若此参数值为16，则表明在谱面开始前添加16个1/8拍的长度）
 #		*谱面开始前：从sav文件的delay表示的时间开始，若delay为0.063，则表示从原音频的63ms处谱面开始，此时在音频前添加的空白长度为16个1/8拍的长度减去63ms
 # -json 1.expert.json 输出的json文件名
 # -outmp3 bgm001.mp3 输出的mp3文件名
+# -skill 32,2,128,4,224,4,320,6 技能节点，最多六个
 def process():
 	argvDict = {}
 	index = 1
+	skills = []
 	while index < len(sys.argv):
 		currArg = sys.argv[index]
 		if currArg == "-sav":
@@ -177,6 +189,20 @@ def process():
 			index += 1
 		elif currArg == "-map":
 			argvDict["map"] = True
+		elif currArg == "-simulator":
+			argvDict["simulator"] = True
+		elif currArg == "-skill":
+			skillStr = sys.argv[index + 1]
+			index += 1
+			# 处理skillStr
+			skillStrList = skillStr.split(",")
+			for i in xrange(len(skillStrList)):
+				if i % 2 == 1:
+					continue
+				skillPos = skillStrList[i]
+				skillLine = int(skillStrList[i + 1])
+				skills.append((skillPos, skillLine))
+			
 		index += 1
 	
 	if "sav" not in argvDict:
@@ -218,6 +244,16 @@ def process():
 		startInst.addLineInst(nodeInst)
 	for nodeInst in lineSList:
 		nodeInst.sortLineInst()
+
+	# 添加skill节点：
+	for skillPos, skillLine in skills:
+		noteInst = noteMap.get(skillPos, {}).get(skillLine)
+		if noteInst is None:
+			print "No note at {}, {}".format(skillPos, skillLine)
+		elif noteInst.__class__.__name__ not in ["LineStart", "NoteN"]:
+			print "Cannot change {} note to skill note.".format(noteInst.__class__.__name__)
+		else:
+			noteInst.setToSkill()
 	
 	# 计算要添加多长时间的空白：
 	preLength = 0
@@ -225,7 +261,7 @@ def process():
 	valid = False
 	if "preLength" in argvDict:
 		preLength = argvDict["preLength"]
-		length = 60.0 / (bpm * 2) * (preLength - 1) # 虽然不知道为什么，但是这里好像要减去1/8拍
+		length = 60.0 / (bpm * 2) * (preLength - 1) # 虽然不知道为什么，但是这里好像要减去1/8拍？暂时存疑，看看这个方法是不是每首歌都适用。
 		if length >= delay:
 			valid = True
 	
@@ -234,7 +270,7 @@ def process():
 		preLengthFloat = delay / (60.0 / (bpm * 2)) / 8
 		preLengthFloat = math.ceil(preLengthFloat)
 		preLength = int(preLengthFloat) * 8
-		length = 60.0 / (bpm * 2) * (preLength - 1) # 虽然不知道为什么，但是这里好像要减去1/8拍
+		length = 60.0 / (bpm * 2) * (preLength - 1) # 虽然不知道为什么，但是这里好像要减去1/8拍？暂时存疑，看看这个方法是不是每首歌都适用。
 		
 	bpmInfo["preLength"] = preLength
 	
@@ -250,15 +286,6 @@ def process():
 	times = noteMap.keys()
 	times.sort(key=lambda item: float(item))
 	
-	mapType = "simulator"
-	if "map" in argvDict:
-		resultList.append({
-			"type": "BPM",
-			"bpm": bpm,
-			"time": 0,
-		})
-		mapType = "chart"
-
 	for tm in times:
 		notes = noteMap[tm]
 		noteList = notes.values()
@@ -274,13 +301,18 @@ def process():
 				"lane": [noteList[0].line, noteList[1].line],
 				"time": calcTime(tm),
 			})
-			
-	import json
-	with open("graphics/{}/{}".format(mapType, argvDict.get("json", "1.expert.json")), "wt") as output:
-		json.dump(resultList, output)
-		
-		
+	
+	if "simulator" in argvDict:
+		with open("graphics/simulator/{}".format(argvDict.get("json", "1.expert.json")), "wt") as output:
+			json.dump(resultList, output)
+	if "map" in argvDict:
+		resultList.insert(0, {
+			"type": "BPM",
+			"bpm": bpm,
+			"time": 0,
+		})
+		with open("graphics/chart/{}".format(argvDict.get("json", "1.expert.json")), "wt") as output:
+			json.dump(resultList, output)
+
 if __name__ == "__main__":
 	process()
-
-
